+++ LibCore/Graphics/GLUtils/IndexBufferObjectSubData.cs
@@ -23,7 +23,6 @@
 // ///////////////////////////////////////////////////////////////////////////////
 
 
-using LibGDXSharp.LibCore.Utils;
 using LibGDXSharp.LibCore.Utils.Buffers;
 
 namespace LibGDXSharp.LibCore.Graphics.GLUtils;
@@ -231,7 +230,7 @@ public class IndexBufferObjectSubData : IIndexData
 
     private int CreateBufferObject()
     {
-        var result = Core.Gdx.GL20.GLGenBuffer();
+        var result = ( int )Core.Gdx.GL20.GLGenBuffer();
 
         Core.Gdx.GL20.GLBindBuffer( IGL20.GL_ELEMENT_ARRAY_BUFFER, result );
         Core.Gdx.GL20.GLBufferData( IGL20.GL_ELEMENT_ARRAY_BUFFER, _byteBuffer.Capacity, null!, _usage );
